% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{BiGaussian}
\alias{BiGaussian}
\title{Bivariate Gaussian}
\usage{
BiGaussian(x, sigma1, sigma2)
}
\arguments{
\item{x}{A vector or matrix with the values where the density derivatives will be calculated.}

\item{sigma1}{The variance of X1}

\item{sigma2}{The variance of X2}
}
\value{
Returnes the  value \eqn{\boldsymbols{\theta}} at \code{x}.
  Eihter as a vector or matrix.
}
\description{
Calculates the true values of a bivarate Gaussian random variable at a grid
of given points
}
