% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{circle_radius}
\alias{circle_radius}
\title{Ridge Radius Circle Data}
\usage{
circle_radius(r = 1, sd = 0.1)
}
\arguments{
\item{r}{Radius used in the generation of the circle data.}

\item{sd}{Standard deviation used in the generation of the circle data.}
}
\value{
Returns the radius of the true ridge underlying the circle data
  generated with radius \code{r} and standard deviation \code{sd}.
}
\description{
Calculates the radius of the true ridge underlying the circle data.
}
\details{
The calculation uses a bisection algorithm.
}
