% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel.R
\name{D2K.Epan}
\alias{D2K.Epan}
\title{2nd derivative of the Epanechnikov Kernel}
\usage{
D2K.Epan(z)
}
\arguments{
\item{z}{A matrix containing on which the kernel is evaluated. If
it is a matrix each column is interpreted as a vector.}
}
\value{
If the input is a vector the vector \eqn{D^{\otimes 2} K(z)} will be
  returned. If it is a matrix a matrix \eqn{D^{\otimes 2}K(z_1), \ldots,
  D^{\otimes 2}K(z_m)} will be returned, where \eqn{z = [z_1, \ldots, z_m]}.
}
\description{
This function is used inside the function \code{D2K} if \code{type =
"Epan"}.
}
