% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel.R
\name{D2K}
\alias{D2K}
\title{2nd order derivative of the kernel function}
\usage{
D2K(z, type = "Gaussian")
}
\arguments{
\item{z}{A vector or matrix containing where the kernel is evaluated. If
it is a matrix each column is interpreted as a vector.}

\item{type}{A character string specify the used kernel function. Current
options are "Gaussian" and...}
}
\value{
If the input is a vector the vector \eqn{D^{\otimes 2} K(z)} will be
  returned. If it is a matrix a matrix \eqn{D^{\otimes 2}K(z_1), \ldots,
  D^{\otimes 2}K(z_m)} will be returned, where \eqn{z = [z_1, \ldots, z_m]}.
}
\description{
\code{D2K} returns the values of the 2nd order derivative of the kernel
function at its argument(s).
}
\examples{
z <- rnorm(3)
D2K(z, "Gaussian")

Z <- matrix(rnorm(12), 3, 4)
D2K(Z, "Gaussian")
}
