% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel.R
\name{DK.Gaussian}
\alias{DK.Gaussian}
\title{1st derivative of the Gaussian Kernel}
\usage{
DK.Gaussian(z)
}
\arguments{
\item{z}{A matrix containing on which the kernel is evaluated. If
it is a matrix each column is interpreted as a vector.}
}
\value{
If the input is avector the value \eqn{DK(z)} will be returned as a
  vector. If the input is a matrix a matrix \eqn{DK(z_1), \ldots, DK(z_m)}
  will be returned, where \eqn{z = [z_1, \ldots, z_m]}.
}
\description{
This function is used inside the functions \code{DK} if \code{type =
"Gaussian"}.
}
