% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel.R
\name{DK}
\alias{DK}
\title{Gradient of the kernel function}
\usage{
DK(z, type = "Gaussian")
}
\arguments{
\item{z}{A vector or matrix containing where the kernel is evaluated. If
it is a matrix each column is interpreted as a vector.}

\item{type}{A character string specify the used kernel function. Current
options are "Gaussian" and...}
}
\value{
If the input is avector the value \eqn{DK(z)} will be returned as a
  vector. If the input is a matrix a matrix \eqn{DK(z_1), \ldots, DK(z_m)}
  will be returned, where \eqn{z = [z_1, \ldots, z_m]}.
}
\description{
\code{DK} returns the gradient of K at its argument(s).
}
\examples{
z <- rnorm(3)
DK(z, "Gaussian")

Z <- matrix(rnorm(12), 3, 4)
DK(Z, "Gaussian")
}
