% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{get_vecId}
\alias{get_vecId}
\title{vec(I_d)}
\usage{
get_vecId(d)
}
\arguments{
\item{d}{Dimension of the vectorized identity matrix.}
}
\value{
A vector of dimension d^2.
}
\description{
Returning the vectorized identity matrix of dimension d.
}
