% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{Grid2D}
\alias{Grid2D}
\title{Grid in 2D}
\usage{
Grid2D(x, y, length = NULL, N = NULL)
}
\arguments{
\item{x}{The interval of the x coordinates}

\item{y}{The interval of the y coordinates}

\item{N}{Number of gridpoints.}

\item{lenght}{Horizontal and vertical distance between two points on the grid.
If only one number is given it will be used for the horizontal and vertical
distance. Ignored if \code{N} is specified}
}
\value{
Returns a 2 x N matrix which contains all the gridpoints.
}
\description{
Creat a matrix which corresponds to a grid of points on the plain
}
