% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{Integrate}
\alias{Integrate}
\title{Stochastic Integration}
\usage{
Integrate(g, d, type = "Gaussian", eps = 1e-04, alpha = 0.01,
  min = 100, max = Inf)
}
\arguments{
\item{g}{A function we want to integrate. Takes a vector as input and returns
a vector.}

\item{d}{Dimension of the input argument of \code{g}.}

\item{type}{the used kernel}

\item{eps}{\code{eps} is the we allow our estimator to vary compared to
the true value with certainty \code{1 - alpha}.}

\item{alpha}{\code{alpha} is the certainty with which our estimator is in a
certained interval around the true value.}

\item{min}{The minimal number of simulations.}

\item{max}{The maximal number of simulations.}
}
\value{
An vector estimating the integral.
}
\description{
Estimates the integral \eqn{\int K(x) g(x) dx} if K is a density function (or
kernel function). \code{espilon, alpha} are control parameters.
}
