% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel.R
\name{K}
\alias{K}
\title{Kernel function}
\usage{
K(z, type = "Gaussian")
}
\arguments{
\item{z}{A vector or matrix containing where the kernel is evaluated. If
it is a matrix each column is interpreted as a vector.}

\item{type}{A character string specify the used kernel function. Current
options are "Gaussian" and...}
}
\value{
If the input is a vector the value \eqn{K(z)} will be returned. If
  the input is a matrix a vector \eqn{K(z_1), \ldots, K(z_m)} will be
  returned, where \eqn{z = [z_1, \ldots, z_m]}.
}
\description{
\code{K} returns the value of the kernel function at its argument(s).
}
\examples{
z <- rnorm(3)
K(z, "Gaussian")

Z <- matrix(rnorm(12), 3, 4)
K(Z, "Gaussian")
}
