% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel.R
\name{K.Uniform}
\alias{K.Uniform}
\title{Uniform Kernel}
\usage{
K.Uniform(z)
}
\arguments{
\item{z}{A matrix containing where the kernel is evaluated. If
it is a matrix each column is interpreted as a vector.}
}
\value{
If the input is a vector the value \eqn{K(z)} will be returned. If
  the input is a matrix a vector \eqn{K(z_1), \ldots, K(z_m)} will be
  returned, where \eqn{z = [z_1, \ldots, z_m]}.
}
\description{
This function is used inside the functions \code{K} if \code{type =
"Uniform"}.
}
