% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalMoments.R
\name{LM0}
\alias{LM0}
\title{Sample local 0-moment}
\usage{
LM0(x, h, type = "Gaussian", data)
}
\arguments{
\item{x}{A vector or matrix containing the points at which the moment will be
calculated.}

\item{h}{A number specifying the bandwidth.}

\item{type}{A character string specifying the used kernel function.}

\item{data}{A data matrix containing in each column a data point.}
}
\value{
If \code{x} is a vector returns the value \eqn{s_n^0(x)}. If the
  input is a matrix a vector \eqn{s_n^0(x_1), \ldots, s_n^0(x_m)} is
  returned, where \eqn{x = [x_1, \ldots, x_m]}.
}
\description{
\code{LM0} returns the value of the sample local 0-moment of the data
points.
}
\examples{
data <- matrix(rnorm(20),2,10)
x <- matrix(c(seq(-2, 2, length.out = 9), rep(0, 9)), 2, 9)
LM0(x, 0.5, "Gaussian", data)
}
