% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalMoments.R
\name{LM0_D2K}
\alias{LM0_D2K}
\title{Sample local 0-moment of \eqn{D^{\otimes 2}K}}
\usage{
LM0_D2K(x, h, type = "Gaussian", data)
}
\arguments{
\item{x}{A vector or matrix containing the points at which the moment will be
calculated.}

\item{h}{A number specifying the bandwidth.}

\item{type}{A character string specifying the used kernel function.}

\item{data}{A data matrix containing in each column a data point.}
}
\value{
If \code{x} is a vector it returns the vector \eqn{\frac{1}{nh^2}
  \sum_{i=1}^n (D^{\otimes 2}K)_h (X_i - x)}. If \code{x} is a matrix it
  returns the matrix \eqn{\frac{1}{nh^2} \sum_{i=1}^n (D^{\otimes 2}K)_h (X_i
  - x_1), \ldots, \frac{1}{nh^2} \sum_{i=1}^n (D^{\otimes 2}K)_h (X_i -
  x_m)}, where \eqn{x = [x_1, \ldots, x_m]}.
}
\description{
Returns the sample local 0-moment of the 2nd order derivatives of the kernel
function.
}
