% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalMoments.R
\name{LM0_DK}
\alias{LM0_DK}
\title{Sample local 0-moment of \eqn{DK}}
\usage{
LM0_DK(x, h, type = "Gaussian", data)
}
\arguments{
\item{x}{A vector or matrix containing the points at which the moment will be
calculated.}

\item{h}{A number specifying the bandwidth.}

\item{type}{A character string specifying the used kernel function.}

\item{data}{A data matrix containing in each column a data point.}
}
\value{
If \code{x} is a vector it returns the vector \eqn{Q_n^{\otimes
  0}(x)}. If \code{x} is a matrix it returns the matrix \eqn{Q_n^{\otimes
  0}(x_1), \ldots, Q_n^{\otimes 0}(x_m)}, where \eqn{x = [x_1, \ldots, x_m]}.
}
\description{
Returns the sample local 0-moment of the 1st order derivative of the kernel
function.
}
