% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalMoments.R
\name{LM1}
\alias{LM1}
\title{Sample local 1-moments}
\usage{
LM1(x, h, type = "Gaussian", data)
}
\arguments{
\item{x}{A vector or matrix containing the points at which the moment will be
calculated.}

\item{h}{A number specifying the bandwidth.}

\item{type}{A character string specifying the used kernel function.}

\item{data}{A data matrix containing in each column a data point.}
}
\value{
If \code{x} is a vector returns the vector \eqn{s_n^{\otimes 1}(x)}. If the
  input is a matrix the matrix \eqn{s_n^{\otimes 1}(x_1), \ldots, s_n^{\otimes 1}(x_m)} is
  returned, where \eqn{x = [x_1, \ldots, x_m]}.
}
\description{
\code{LM1} returns a vector with the 1st order sample local moments.
}
\examples{
data <- matrix(rnorm(20),2,10)
x <- matrix(c(seq(-2, 2, length.out = 9), rep(0, 9)), 2, 9)
LM1(x, 0.5, "Gaussian", data)
}
