% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{LogDensity}
\alias{LogDensity}
\title{Local Estimator of the log-density}
\usage{
LogDensity(x, h, type = "Gaussian", method = "LHSE", data)
}
\arguments{
\item{x}{A vector or matrix containing the points at which the estimator will be
calculated.}

\item{h}{A number specifying the bandwidth.}

\item{type}{A character string specifying the used kernel function.}

\item{method}{A cherecter string specifying the uesed method. The possible
methods are \code{LHSE, LLLE, LMME, KDE}.}

\item{data}{A data matrix containing in each column a data point.}
}
\value{
Returnes the the value \eqn{\hat{\boldsymbols{\theta}}} at \code{x}.
  Eihter as a vector or matrix.
}
\description{
Returns the local estimator of \eqn{\hat{\boldsymbols{\theta}}} for the
location(s) \eqn{x}
}
