% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{LogEstimation}
\alias{LogEstimation}
\title{Log-derivative transformation}
\usage{
LogEstimation(est)
}
\arguments{
\item{est}{Estimator of the density derivatives in R^{1 + d + d^2}.}
}
\value{
Estimator of the log-density derivatives in R^{d + d^2}.
}
\description{
Transforms the vector of density derivatives up to second order into a vector
of log-density derivatives of 1st and 2nd order.
}
