% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{LSME}
\alias{LSME}
\title{Local Estimator of the log-density with LSME}
\usage{
LSME(x, h, type = "Gaussian", data)
}
\arguments{
\item{x}{A vector or matrix containing the points at which the moment will be
calculated.}

\item{h}{A number specifying the bandwidth.}

\item{type}{A character string specifying the used kernel function.}

\item{data}{A data matrix containing in each column a data point.}
}
\value{
Returnes a list with
\describe{
\item{theta}{Dataframe containig the evaluated points and estimators.}
\item{method}{Used method.}
\item{h}{Used bandwidth.}
\item{type}{Used kernel type.}
\item{data}{Data [d, n].}}
}
\description{
Returns the local estimator of \eqn{\hat{\boldsymbols{\theta}}} for the
location(s) \eqn{x} calculated with the LSME
}
