% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{MatToDf}
\alias{MatToDf}
\title{Matrix to data frame}
\usage{
MatToDf(data)
}
\arguments{
\item{data}{A matrix where each column represents a data point in R^d}
}
\value{
A data frame where each row represents a data point and each column
  contains one component of each data point. The colums are named with X1, X2,...
}
\description{
Transforms the data matrix into a data frame usealbe by ggplot 2
}
