% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanShiftAlgoritms.R
\name{ModeSeeking}
\alias{ModeSeeking}
\title{Mode Seeking Algorithm}
\usage{
ModeSeeking(x, h, type = "Gaussian", method = "LHSE", data,
  delta = 1e-10)
}
\arguments{
\item{x}{Starting point [d]}

\item{h}{Kernel bandwidth parameter}

\item{type}{Used kernel. Either \code{Gaussian} or \code{Epan}-(echikov).}

\item{method}{Used method for the log-density estimation. Either \code{LHSE},
\code{LMME} or \code{KDE}.}

\item{data}{Data matrix [d, n].}

\item{delta}{Stopping criteria.}
}
\value{
Fixpoints of the iteration [d]
}
\description{
Gradient acent with step size correction according to the direction of the gradient
}
