% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{PlotData}
\alias{PlotData}
\title{Plot raw data}
\usage{
PlotData(data, ...)
}
\arguments{
\item{data}{A matrix containing in each column a data point or a Dataframe}
}
\value{
Plots the raw data.
}
\description{
Plots the raw data projected in 2 dimensions.
}
