% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanShiftAlgoritms.R
\name{RidgeSeeking}
\alias{RidgeSeeking}
\title{Ridge Seeking Algorithm}
\usage{
RidgeSeeking(x, h, type = "Gaussian", method = "LHSE", data,
  delta = 1e-06, iter_max = Inf, path = FALSE)
}
\arguments{
\item{x}{Starting point [d]}

\item{h}{Kernel bandwidth parameter}

\item{type}{Used kernel. Either \code{Gaussian} or \code{Epan}-(echikov).}

\item{method}{Used method for the log-density estimation. Either \code{LHSE},
\code{LMME} or \code{KDE}.}

\item{data}{Data matrix [d, n].}

\item{delta}{Stopping criteria.}

\item{iter_max}{Maximal number of iterations before the procedure stops. Will
print a warning if it did not converge.}

\item{path}{\code{TRUE} if the whole path of the iteration should be returned.}
}
\value{
Fixpoints of the iteration as vector [d]. If \code{path = TRUE} it returns a list with
\describe{\item{x}{Fixpoint [d] of the iteration.}
\item{path}{Path of the iteration [d, iter], where iter is the number of steps.}}
}
\description{
Gradient acent with step size correction according to the direction of the gradient
}
