% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanShiftAlgoritms.R
\name{SubspaceProjection}
\alias{SubspaceProjection}
\title{Subspace Projection}
\usage{
SubspaceProjection(x, theta2)
}
\arguments{
\item{x}{Vector [d]}

\item{theta2}{The vectorized Hessian of the log-density [d^2]}
}
\value{
Projects a vector on the d-1 dimensional subspace spanned by the d-1
eigenvectors corresponding to the d-1 smallest eigenvalues of the Hessian of
the log-density.
}
\description{
Subspace Projection
}
