% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{Test.KDE}
\alias{Test.KDE}
\title{Test.KDE}
\usage{
Test.KDE(theta.LHSE, h, var1, var2)
}
\arguments{
\item{h}{bandwith used for the estimators}

\item{var1}{Variance on X1}

\item{var2}{Variance on X2}

\item{tehta.LHSE}{Dataframe with the estimators of the LHSE}
}
\description{
Testfunction comparing LHSE and KDE with Gauss data. The output is the
adjusted LHSE (with the bias of the KDE), which should now fit the KDE up to
order h^2.
}
