% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge.R
\name{test.ridge}
\alias{test.ridge}
\title{Test of ridge point}
\usage{
test.ridge(est, x, n, h, type, alpha = 0.05, S.tilde)
}
\arguments{
\item{est}{Vector with the estimator of the log-density derivatives.}

\item{x}{A vector or matrix containing the points at which the test will be
executed.}

\item{n}{Sample size}

\item{h}{A number specifying the bandwidth.}

\item{type}{A character string specifying the used kernel function.}

\item{alpha}{Confidence level}

\item{S.tilde}{Hessian matrix of the sample score}
}
\value{
Returns a logical wheter or not \code{x} is a ridge point or not.
}
\description{
\code{test.ridge} returns TRUE if with confidence 1-alpha there is a ridge
point at x.
}
